requirements.txt
pymongo==10.10.10.10
python-telegram-bot==21.10
google-generativeai==0.8.4
google_search_results==2.4.2
python-dotenv==1.0.1
pillow==10.2.0
spacy==3.8.4
serpapi==0.1.5
transformers==4.48.1
requests==2.31.0